+++ lab1/kiennam99/validate.sh
